+++ python/paddle/fluid/layers/__init__.py
@@ -18,11 +18,9 @@ from . import io
 from .io import *
 from . import math_op_patch
 from .math_op_patch import *
-from .learning_rate_scheduler import *
 from ..layer_helper import LayerHelper
 
 
 __all__ = []
 __all__ += nn.__all__
 __all__ += io.__all__
-__all__ += learning_rate_scheduler.__all__
