+++ python/paddle/fluid/layers/nn.py
@@ -54,61 +54,3 @@ from ..data_feeder import (
 from paddle.utils import deprecated
 from paddle import _C_ops, _legacy_C_ops
 from collections.abc import Iterable
-
-
-__all__ = [
-    'autoincreased_step_counter',
-]
-
-
-def autoincreased_step_counter(counter_name=None, begin=1, step=1):
-    """
-    :api_attr: Static Graph
-
-    Create an auto-increase variable. which will be automatically increased
-    by 1 in every iteration. By default, the first return of this counter is 1,
-    and the step size is 1.
-
-    Args:
-        counter_name(str, optional): The counter name. Default '@STEP_COUNTER@'.
-        begin(int, optional): The first return value of this counter. Default 1.
-        step(int, optional): The step size. Default 1.
-
-    Returns:
-        Variable: The auto-increased Variable with data type int64.
-
-    Examples:
-        .. code-block:: python
-
-           import paddle.fluid as fluid
-           import paddle
-           paddle.enable_static()
-           global_step = fluid.layers.autoincreased_step_counter(
-               counter_name='@LR_DECAY_COUNTER@', begin=0, step=1)
-    """
-    helper = LayerHelper('global_step_counter')
-    if counter_name is None:
-        counter_name = '@STEP_COUNTER@'
-    counter, is_new_var = helper.create_or_get_global_variable(
-        name=counter_name,
-        dtype='int64',
-        shape=[1],
-        persistable=True,
-        belong_to_optimizer=True,
-    )
-    if is_new_var:
-        helper.set_variable_initializer(
-            counter,
-            initializer=paddle.nn.initializer.ConstantInitializer(
-                value=begin - 1, force_cpu=True
-            ),
-        )
-        helper.main_program.global_block()._prepend_op(
-            type='increment',
-            inputs={'X': [counter]},
-            outputs={'Out': [counter]},
-            attrs={'step': float(step)},
-        )
-        counter.stop_gradient = True
-
-    return counter
