+++ python/paddle/static/__init__.py
@@ -72,8 +72,8 @@ from .nn.control_flow import Print  # noqa: F401
 from ..fluid.param_attr import WeightNormParamAttr  # noqa: F401
 from ..fluid.optimizer import Optimizer  # noqa: F401
 
-from ..fluid.layers import exponential_decay  # noqa: F401
-from ..fluid.layers import learning_rate_scheduler  # noqa: F401
+from ..optimizer.lr import exponential_decay  # noqa: F401
+from ..optimizer import lr  # noqa: F401
 
 from .nn.metric import auc  # noqa: F401
 from .nn.metric import accuracy  # noqa: F401
